old-requirements.txt
# Core Dependencies
asyncio
python-dotenv==0.21.0
pytest==7.4.4
pytest-asyncio==0.23.3

# LangGraph and LangChain (Latest Versions)
langgraph==1.0.2
langgraph-cli[inmem]==0.4.3
langchain==0.3.21
langchain-core==0.3.42
langchain_google_genai==2.1.9
langchain-mcp-adapters==0.2.0
langchain-tavily==0.1.2
langchain_community==0.3.20

# Google AI Integration
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.14

# MCP Support
mcp==1.2.0

# Image Processing
Pillow==10.1.0
pytesseract==0.3.10

# Document Processing
docling==1.31.0

# Code Quality
ruff==0.7.3